Getting Started Guide - Completed
Finished by connecting the model, controller, and view of articles
together.
Added a comment model with a relationship to an article entry, then
connected it to it's respective controller and views.
Partials were created for reusable content and are represented by using
'_' before the file name '_form.html.erb', for example.

Next step, Active Record Basics